New one

+++ app/src/main/java/com/appli/ilink/HelpActivity.java
@@ -6,6 +6,6 @@ import android.support.v7.app.AppCompatActivity;
 public class HelpActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView((int) R.layout.activity_help);
+        setContentView(R.layout.activity_help);
     }
 }
