+++ app/src/main/java/com/appli/ilink/ModifyPasswordActivity.java
@@ -6,6 +6,6 @@ import android.support.v7.app.AppCompatActivity;
 public class ModifyPasswordActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView((int) R.layout.activity_modify_password);
+        setContentView(R.layout.activity_modify_password);
     }
 }
